feature: support arm gicv2
This commit adds support for version 2 of the ARM Generic Interrupt
Controller. The interface between version 2 and 3/4 was largely kept
the same. The largest differences technically is the absent system
register interface as well as the absent third peripheral for
handling software-generated interrupts.

To compensate for the physically different hardware, minor modifications
to the gic interface were necessary (for ex. processor selection),
making it necessary to move the gicv2 into it's own module.

Signed-off-by: Fritz Stracke <fritz.stracke@rwth-aachen.de>